Triangulation stores data associated with sites
Each site of the triangulation is now associated with a data object.
Therefore the Triangulation now has a generic type parameter that
defines the type of objects that can be inserted into the triangulation
along with the sites. The delaunayPlace() method accepts a data object
in addition to the site parameter for insertion.